Nearly-done docker setup
only took three hours and it still doesn't do everything it ought to,
kill me now.
Seriously: re-generating on changes does not work properly, but this
could also be a pelican issue (or a firefox issue or something else).
But probably it's just me holding docker volumes the wrong way.